sim: Break out sim file reading into simfs

+++ include/sim.h
@@ -204,7 +204,7 @@ int ofono_sim_write(struct ofono_sim *sim, int id,
 			const unsigned char *data, int length, void *userdata);
 
 int ofono_sim_read_bytes(struct ofono_sim *sim, int id,
-			unsigned short offset, int num_bytes,
+			unsigned short offset, unsigned short num_bytes,
 			ofono_sim_file_read_cb_t cb, void *data);
 #ifdef __cplusplus
 }
